Refactor studio image upload functionality and remove StudioPortfolioForm component
